[IMP] improving the time and memory consumption
The process originally used around 4GiB of RAM
to insert batches of 1M rows and took 12min 27seg
to complete the import process. By changing the
chunk size by 10K, we gain a little speed boost
that makes the process take 12min 06seg to finish
and also uses 93MiB, which is in fact, a huge
optimization.